refactor: mutation, query refactored
