+++ hooks/query/useAuthQuery.ts
@@ -1,7 +1,6 @@
 import { signupApi } from "@/api/auth/api";
 import { UseQueryOptions, useQuery } from "react-query";
 import { DefaultResponse, FailureResponse } from "types/api";
-import { SignupRequestType } from "./useReviewQuery";
 
 export const signupQueryKey = () => ["signup"];
