+++ interface/types.ts
@@ -2,10 +2,6 @@ import { UpdateReviewRequestTypeV2 } from "@/api/review/typedef";
 
 export type AllKeyboardEvent = React.KeyboardEvent | KeyboardEvent;
 
-// export interface CategoryFilterListType {
-//   [index: string]: string[];
-// }
-
 export interface CategoryFilterListType {
   title: CategoryTitle;
   tags: (string | number)[];
@@ -23,13 +19,6 @@ export interface BeerFilterType {
   volume_min: number;
   volume_max: number;
 }
-
-// export enum CategoryTitle {
-//   CRITERIA = "정렬 기준",
-//   SORT = "맥주 종류",
-//   COUNTRY = "제조국",
-//   ALCHOLE = "도수",
-// }
 
 export enum CategoryTitle {
   SORT_CRITERIA = "정렬 기준",
@@ -87,12 +76,6 @@ export type ErrorResponse = {
   timestamp: string;
 };
 
-export interface ReviewInfoType extends UpdateReviewRequestTypeV2 {
-  beerName?: string;
-}
-
-export type UpdatedReviewInfo = Omit<ReviewInfoType, "beerName">;
-
 export enum ReviewSortEnum {
   RecentlyUpdated = "RECENTLY_UPDATED",
   MostLikes = "MOST_LIKES",
@@ -131,7 +114,7 @@ export enum OAUTH_PROVIDER {
 }
 
 export interface ReviewPostparamType {
-  content?: string;
+  content: string;
   rate: number;
   image_url?: string[];
   buy_from?: string[];
